requirements.txt
django==5.2.4
environs~=14.2.0
psycopg2-binary==2.9.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
setuptools==80.9.0